[hack] Make normalize_inter more efficient
Summary: The current implementation is O(len(tyl1) * len(tyl2)), which scales
pretty badly. In particular, we run into this problem when typing large arrays.
Hack around it by switching to plain list append once the lists cross a certain
threshold, which is linear in time, at the expense of using more memory.

Reviewed By: @dabek, @dlreeves

Differential Revision: D2325674