Use typeof consistently
Summary: Since typeof is an operator and not a function it could be confusing for some users to see it used as if it were a function.

Additionally, since all other typeof usage in esprima_test_runner.js was using typeof without parenthesis, I figured making it consistent would be good.
Closes https://github.com/facebook/flow/pull/690

Reviewed By: @mroch

Differential Revision: D2318235

Pulled By: @gabelevi